feat: add option to split kwargs and rename module
This change is to facilitate integration with pycbc inference.
